+ Add: ActorTab.cs and ClassTab.cs re-commit
- Modified: ActorTab.cs refactoring
- Modified: Refactoring DatabaseMain.cs
